Fix blank screen
When steam is installed, the windows are black, without content.
But when the snap is refreshed, that is fixed. The reason is
that the ld.so.conf file is refreshed in the post-refresh hook,
and so the code can't find the Gtk3 libraries.

This MR fixes this by moving the script from the `post-refresh`
to the `configure` hook.