Distribute the whole i386 libraries
The snap store has a limit of 30 elements in the Layout section.
After the last patch, the number was bigger, so this MR fixes
this. To do so, it:

- removes all the individual i386 bindings and does a single
bind for /usr/lib/i386-linux-gnu (this implies to also copy
all the i386 files from Core22)

- removes unused entries; specifically the mesa and mesa-egl
bindings, since those folders don't exist anymore, neither in
the x86_64 section, nor in the i386 section.

It also fixes the $SNAP/lib/steam bind, pointing it to
$SNAP/usr/lib/steam.